+++ types/index.d.ts
@@ -112,7 +112,7 @@ export declare class AeroflyMission {
      */
     isFeatured: boolean;
     /**
-     * @property {number|undefined} difficulty in 0..2, percentage
+     * @property {number|undefined} difficulty values between 0.00 and 2.00 have been encountered, but they seem to be without limit
      */
     difficulty: number | undefined;
     /**
@@ -144,6 +144,10 @@ export declare class AeroflyMission {
      */
     duration: number;
     /**
+     * @property {?AeroflyMissionTargetPlane} finish as finish condition
+     */
+    finish: AeroflyMissionTargetPlane | null;
+    /**
      * @property {AeroflyMissionConditions} conditions like time and weather for mission
      */
     conditions: AeroflyMissionConditions;
@@ -158,7 +162,7 @@ export declare class AeroflyMission {
      * @param {AeroflyLocalizedText[]} [additionalAttributes.localizedTexts] translations for title and description
      * @param {string[]} [additionalAttributes.tags]
      * @param {boolean} [additionalAttributes.isFeatured] makes this mission pop up in "Challenges"
-     * @param {number|undefined} [additionalAttributes.difficulty] 0..1, percentage
+     * @param {number|undefined} [additionalAttributes.difficulty] values between 0.00 and 2.00 have been encountered, but they seem to be without limit
      * @param {"taxi"|"takeoff"|"cruise"|"approach"|"landing"|"winch_launch"|"aerotow"} [additionalAttributes.flightSetting] of aircraft, like "taxi", "cruise"
      * @param {{name:string,livery:string,icao:string}} [additionalAttributes.aircraft] for this mission
      * @param {string} [additionalAttributes.callsign] of aircraft, uppercased
@@ -166,6 +170,7 @@ export declare class AeroflyMission {
      * @param {object} [additionalAttributes.destination] position of aircraft, as well as name of destination airport. Position does not have match airport.
      * @param {number} [additionalAttributes.distance] in meters
      * @param {number} [additionalAttributes.duration] in seconds
+     * @param {?AeroflyMissionTargetPlane} [additionalAttributes.finish] as finish condition
      * @param {AeroflyMissionConditions} [additionalAttributes.conditions] like time and weather for mission
      * @param {AeroflyMissionCheckpoint[]} [additionalAttributes.checkpoints] form the actual flight plan
      */
@@ -184,6 +189,7 @@ export declare class AeroflyMission {
             destination,
             distance,
             duration,
+            finish,
             conditions,
             checkpoints,
         }?: {
@@ -199,6 +205,7 @@ export declare class AeroflyMission {
             destination?: AeroflyMissionPosition;
             distance?: number;
             duration?: number;
+            finish?: AeroflyMissionTargetPlane | null;
             conditions?: AeroflyMissionConditions;
             checkpoints?: AeroflyMissionCheckpoint[];
         },
@@ -418,9 +425,9 @@ export declare class AeroflyMissionCheckpoint {
      */
     frequency: number | null;
     /**
-     * @property {boolean} flyOver if waypoint is meant to be flown over
+     * @property {boolean|undefined} flyOver if waypoint is meant to be flown over
      */
-    flyOver: boolean;
+    flyOver: boolean | undefined;
     /**
      * @param {string} name ICAO code for airport, runway designator, navaid
      *    designator, fix name, or custom name
@@ -441,7 +448,7 @@ export declare class AeroflyMissionCheckpoint {
      * @param {number?} [additionalAttributes.length] of runway, in meters
      * @param {number?} [additionalAttributes.length_feet] of runway, in feet. Will overwrite length
      * @param {number?} [additionalAttributes.frequency] of runways or navigational aids, in Hz; multiply by 1000 for kHz, 1_000_000 for MHz
-     * @param {boolean} [additionalAttributes.flyOver] if waypoint is meant to be flown over
+     * @param {boolean|undefined} [additionalAttributes.flyOver] if waypoint is meant to be flown over
      */
     constructor(
         name: string,
@@ -465,7 +472,7 @@ export declare class AeroflyMissionCheckpoint {
             length?: number | null;
             length_feet?: number | null;
             frequency?: number | null;
-            flyOver?: boolean;
+            flyOver?: boolean | undefined;
         },
     );
     /**
@@ -541,7 +548,42 @@ export declare class AeroflyLocalizedText {
      * @param {number} index if used in an array will se the array index
      * @returns {string} to use in Aerofly FS4's `custom_missions_user.tmc`
      */
-    toString(index: number): string;
+    toString(index?: number): string;
+}
+export declare class AeroflyMissionTargetPlane {
+    /**
+     * @property {number} longitude easting, using the World Geodetic
+     *    System 1984 (WGS 84) [WGS84] datum, with longitude and latitude units
+     *    of decimal degrees; -180..180
+     */
+    longitude: number;
+    /**
+     * @property {number} latitude northing, using the World Geodetic
+     *    System 1984 (WGS 84) [WGS84] datum, with longitude and latitude units
+     *    of decimal degrees; -90..90
+     */
+    latitude: number;
+    /**
+     * @property {number} dir in degree
+     */
+    dir: number;
+    /**
+     * @property {string} name of property
+     */
+    name: string;
+    /**
+     *
+     * @param {number} longitude easting, using the World Geodetic
+     *    System 1984 (WGS 84) [WGS84] datum, with longitude and latitude units
+     *    of decimal degrees; -180..180
+     * @param {number}latitude northing, using the World Geodetic
+     *    System 1984 (WGS 84) [WGS84] datum, with longitude and latitude units
+     *    of decimal degrees; -90..90
+     * @param {number} dir in degree
+     * @param {string} name of property
+     */
+    constructor(longitude: number, latitude: number, dir: number, name?: string);
+    toString(): string;
 }
 declare const _default: {
     AeroflyMissionsList: typeof AeroflyMissionsList;
@@ -550,6 +592,7 @@ declare const _default: {
     AeroflyMissionConditionsCloud: typeof AeroflyMissionConditionsCloud;
     AeroflyMissionCheckpoint: typeof AeroflyMissionCheckpoint;
     AeroflyLocalizedText: typeof AeroflyLocalizedText;
+    AeroflyMissionTargetPlane: typeof AeroflyMissionTargetPlane;
 };
 export default _default;
 //# sourceMappingURL=index.d.ts.map